requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pandas==2.1.3
scikit-learn>=1.5.0
joblib==1.3.2
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
python-multipart==0.0.6
numpy>=1.25.2,<2.0
scipy>=1.11.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
cryptography>=3.4.0
python-dateutil>=2.8.2
pytz>=2020.1
tzdata>=2022.1
supabase==2.3.4
python-dotenv==1.0.0
